taehan/a2a/requirements.txt
# Core framework packages
a2a-sdk==0.2.8
click==8.2.1
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
uvicorn==0.35.0
numpy==1.26.4

# LangChain ecosystem
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.69
langchain-google-genai==2.1.8
langchain-ollama==0.3.5
langchain-openai==0.3.28
langchain-upstage==0.6.0
langgraph==0.5.3

# Web and data processing
beautifulsoup4==4.13.4
gradio==5.38.0
requests==2.32.4

# Data and database
pandas==2.3.1
sqlalchemy==2.0.41
faiss-gpu==1.7.2

# Utilities
holidays==0.74
protobuf==6.31.1
termcolor==3.1.0
kiwipiepy==0.21.0
rank-bm25==0.2.2
google-adk==1.3.0
